Major restructuring commit to improve modularity
The orginial source code is left unchanged except for import statements that needed correction. However, the files are now structured
in seperate packages (e.g. robots, controllers, simulation with their own setup, README and LICENSE files) that are going to be
installed within a virtual environment (as a next step).

Robot urdf files and meshes are moved to the top level as they are not really 'source' code.