Create Simulation class and refactor World class
The simulation class is now the high-level object that launches the physics server and that should step the simulation
(not implemented like that yet).

The world object is now on its own and does not include the robot or steps the simulation. Additional functions to add
and remove objects from the world are implemented.

The control demos are adapted to the changes and still have the same behaviour as before.