Refactoring of the BulletRobot class
The BulletRobot class is split into two classes, BulletRobotDescription and BulletRobot.
The former loads all relevant information about the robot from the urdf and stores this information
when the object is created, wheareas the latter implements methods that are required during the
actual simulation, e.g. getting the robot state, set the control method, and set the joint
commands, as well as a few saftey methods for checking joint limits, enabling/disabling the
robot, and triggering an emergency stop.

These methods are tested in robot_test.py.

The robot specfic PandaArm class loses importance because most methods are implemented in
BulletRobot and BulletRobotDescription. This class might be replaced with a simple config file in
the future. Or, it might regain importance if we see that different types of robots do require
different methods and classes. It is for now left there to be backwards compatible with the demo
scripts. However, technically, this class is not needed for the simulation of a panda robot.