Add example script to run a simulation with velocity control
The velocity_control_test.py script shows how the simulation is supposed to be structured.
First, the interface, simulation, world, and robot are created and set to the desired state.
Then, the simulation is waiting to accept commands over the interface and applies those until
the connection is broken.

This certainly works for joint position and velocity control. Torque control does not work yet
but will be the very next step.