Refactor robot module (#6)
* Refactoring of the BulletRobot class

The BulletRobot class is split into two classes, BulletRobotDescription and BulletRobot.
The former loads all relevant information about the robot from the urdf and stores this information
when the object is created, wheareas the latter implements methods that are required during the
actual simulation, e.g. getting the robot state, set the control method, and set the joint
commands, as well as a few saftey methods for checking joint limits, enabling/disabling the
robot, and triggering an emergency stop.

These methods are tested in robot_test.py.

The robot specfic PandaArm class loses importance because most methods are implemented in
BulletRobot and BulletRobotDescription. This class might be replaced with a simple config file in
the future. Or, it might regain importance if we see that different types of robots do require
different methods and classes. It is for now left there to be backwards compatible with the demo
scripts. However, technically, this class is not needed for the simulation of a panda robot.

* Make necessary changes to adapt to bullet_robot refactoring

These changes ensure backward compatibility with the demos from the original code.
It's useful to check that these demos still have the same behavior to ensure that the
refactoring isn't a breaking change.

However, it's important to note that these are minimal changes and not all the available
methods are considered, so it might happen that unused methods break the code. This is
acceptable given the fact that this is a development process and that the demos use the
simualtion in a slightly different way that it is intended here.

* Adapt interface test to changes

* Remove private methods from the class header doc and a few docs

* Add example script to run a simulation with velocity control

The velocity_control_test.py script shows how the simulation is supposed to be structured.
First, the interface, simulation, world, and robot are created and set to the desired state.
Then, the simulation is waiting to accept commands over the interface and applies those until
the connection is broken.

This certainly works for joint position and velocity control. Torque control does not work yet
but will be the very next step.

* further improvements